refactor: 클래스의 역할에 맞는 패키지로 이동하라
