+++ src/main/java/lotto/domain/LottoBuyer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.math.BigInteger;
