+++ src/main/java/lotto/domain/LottoConstants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class LottoConstants {
     public static final int COUNT_OF_LOTTO_NUMBERS = 6;
